Update.
	* locale/programs/ld-ctype.c (ctype_finish): Take all characters from
	the input charset into account when generating the hash table.
	(allocate_arrays): Correct setting default width.  Not all empty slots
	in the table are filled, only those not covert explicitly by the
	locale description and in the charset.

	* stdio-common/vfscanf.c: Make sure to always return WEOF and EOF for
	wide character version.
	For %C handling, test correct pointer variable for NULL.

	* wcsmbs/wctob.c: Handle WEOF special.

	* wcsmbs/wcwidth.h: 0xff in width array means invalid character.

	* wctype/wctype.h: Protect gcc-isms with __extension__.  Avoid
	always-true test to avoid warning.

+++ localedata/tests-mbwc/tst_wctrans.c
@@ -27,7 +27,8 @@ tst_wctrans (FILE * fp, int debug_flg)
 
       if (debug_flg)
 	{
-	  fprintf (stderr, "tst_wctrans : [ %d ] ret = %d\n", rec + 1, ret);
+	  fprintf (stderr, "tst_wctrans : [ %d ] ret = %ld\n", rec + 1,
+		   (long int) ret);
 	  fprintf (stderr, "		       errno = %d\n", errno_save);
 	}
